+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - use registry to resolve Chain address
 - support numeric first class data
 - cryptocompare API support
+
+### Changed
+- use `@umb-network/toolbox` for fetching ABIs for contracts